Add files via upload
This code defines a SparseMatrix class to manage the list and a Node struct to represent elements in the linked list. The linked list can be printed in the required format by using the show method once non-zero members have been inserted using the insert method. The primary function shows how to display the linked list as instructed and add elements.